飛行試験

+++ gpio_test/PWM_01.py
@@ -22,7 +22,7 @@ time.sleep(10)
 #duty比を線形に0まで落とす
 while duty_ratio > 0:
     duty_ratio = duty_ratio - 0.01
-    pi.hardware_PWM(gpio_pin0, 2, duty_ratio * 1000000)
+    pi.hardware_PWM(gpio_pin0, 2,  int(duty_ratio * 1000000))
     time.sleep(0.1)
 
 pi.set_mode(gpio_pin0, pigpio.INPUT)
